+++ packages/core-typings/src/IOAuthApps.ts
@@ -3,7 +3,7 @@ export interface IOAuthApps {
 	name: string;
 	active: boolean;
 	clientId: string;
-	clientSecret: string;
+	clientSecret?: string;
 	redirectUri: string;
 	_createdAt: Date;
 	_createdBy: {