+++ packages/livechat/src/routes/ChatFinished/stories.tsx
@@ -2,8 +2,8 @@ import { action } from '@storybook/addon-actions';
 import type { Meta, StoryFn } from '@storybook/preact';
 import type { ComponentProps } from 'preact';
 
-import { loremIpsum, screenDecorator } from '../../../.storybook/helpers';
 import ChatFinished from './component';
+import { loremIpsum, screenDecorator } from '../../../.storybook/helpers';
 
 export default {
 	title: 'Routes/ChatFinished',