+++ packages/livechat/src/routes/GDPRAgreement/stories.tsx
@@ -2,8 +2,8 @@ import { action } from '@storybook/addon-actions';
 import type { Meta, StoryFn } from '@storybook/preact';
 import type { ComponentProps } from 'preact';
 
-import { screenDecorator } from '../../../.storybook/helpers';
 import GDPRAgreement from './component';
+import { screenDecorator } from '../../../.storybook/helpers';
 
 export default {
 	title: 'Routes/GDPRAgreement',