+++ packages/ui-contexts/src/UserContext.ts
@@ -24,6 +24,8 @@ export type Sort<TSchema extends Document = Document> = Exclude<MongoFindOptions
 export type FindOptions<TSchema extends Document = Document> = {
 	fields?: Fields<TSchema>;
 	sort?: Sort<TSchema>;
+	skip?: number;
+	limit?: number;
 };
 
 export type UserContextValue = {
@@ -35,8 +37,6 @@ export type UserContextValue = {
 	) => [subscribe: (onStoreChange: () => void) => () => void, getSnapshot: () => T | undefined];
 	querySubscription: (
 		query: Filter<Pick<ISubscription, 'rid' | 'name'>>,
-		fields?: MongoFindOptions<ISubscription>['projection'],
-		sort?: MongoFindOptions<ISubscription>['sort'],
 	) => [subscribe: (onStoreChange: () => void) => () => void, getSnapshot: () => ISubscription | undefined];
 	queryRoom: (
 		query: Filter<Pick<IRoom, '_id'>>,